Lots of cleanup and moving duplicate functions to the workflow_utilities.py file
