+++ workflow/scrape_paper_info.py
@@ -1,6 +1,16 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+"""
+This script reads in a CSV with into about papers we want in our primary CSV database.
+It checks the primary CSV file to see if a record for that paper already exists.
+If the record is already there and complete, nothing is done.
+If the record is already there but the record is not complete, the script tries to scrape the
+  site to get more information.
+If the record is not there for this paper, it scrapes and creates a new record in the primary CSV database
+"""
+
 import ast
-import traceback
-import time
 import argparse
 import sys
 import datetime
@@ -8,30 +18,30 @@ from pathlib import Path
 import shutil
 
 from dotenv import load_dotenv
-
 import pandas as pd
 
-from .column_definitions import standard_columns, key_mapping
+from column_definitions import standard_columns
 
-from .get_paper_info import get_paper_info, which_literature_site
-
-from .workflow_utilities import labels_fix, abstract_fix, extract, filter_by_lit_site, filter_by_count, save_status
+from workflow_utilities import extract, filter_by_lit_site, filter_by_count, save_status, scrape_paper_info
 
 
-def raw_data_check(df):
-    unlabeled_papers = df[df['label_level_1'].isna()]['url']
+def raw_data_check(df_raw):
+    """
+    Can use this to check the input raw data but not using now
+    """
+    unlabeled_papers = df_raw[df_raw['label_level_1'].isna()]['url']
     print(f'There are {len(unlabeled_papers)} papers without labels')
 
-    duplicate_papers = df[df.duplicated(['url'])]['url'].drop_duplicates().sort_values()
+    duplicate_papers = df_raw[df_raw.duplicated(['url'])]['url'].drop_duplicates().sort_values()
     print(f'There are {len(duplicate_papers)} duplicate papers')
     print(duplicate_papers)
 
-    print(df.duplicated(subset='url', keep='first').sum())
+    print(df_raw.duplicated(subset='url', keep='first').sum())
 
     # Look for commas in the labels
     # Loop through all the records
     papers_with_commas_in_labels = set()
-    for index, row in df[['url', 'label_level_1']].iterrows():
+    for index, row in df_raw[['url', 'label_level_1']].iterrows():
         if not isinstance(row['label_level_1'],float):
             labels_as_string = row['label_level_1']
             labels = ast.literal_eval(labels_as_string)
@@ -46,186 +56,99 @@ def raw_data_check(df):
         print("**** *****\n")
 
 
-def transform(df):
-    # Make an empty table for the results of the transform
-    transformed_df = pd.DataFrame(columns=['title', 'doi', 'abstract', 'labels', 'url',
-                                           'literature_site',
-                                           'full_doc_link', 'is_open_access'])
-
-    # Need to keep track of the status of each attempt to get paper info
-    status_df = pd.DataFrame(columns=['url', 'literature_site', 'get_paper_info_result',
-                                      'title_len', 'abstract_len', 'doi_len',
-                                      'pdf_len', 'is_open_access',
-                                      'num_labels',
-                                      'error_traceback', 'scrape_time'])
-    status_df.astype(int)  # No floats
-
-    # Loop through the records to get paper info
-    for index, row in df[['Primary lit site', 'Functions Level I','Abstract']].iterrows():
-        url, labels, abstract = row
-        # if pd.isnull(labels):
-        #     labels = ''
-        print(f"{index} url: {url}")
-
-        start_time = time.time()
-
-        literature_site = which_literature_site(url)
-
-        # continue
-        title = doi = abstract = full_doc_link = ''
-        is_open_access = False
-
-        # fix labels
-        labels = labels_fix(labels)
-
-        try:
-            paper_info = get_paper_info(url)
-            if paper_info:
-                title, doi, abstract, full_doc_link, is_open_access, is_blocked = paper_info
-                if is_blocked:
-                    get_paper_info_result = 'blocked'
-                else:
-                    get_paper_info_result = 'no_exception'
-
-                    # fix abstract
-                    abstract = abstract_fix(abstract)
-
-                    transformed_df = transformed_df.append({
-                        'title': title,
-                        'doi': doi,
-                        'abstract': abstract,
-                        'labels': labels,
-                        'url': url,
-                        'literature_site': literature_site,
-                        'full_doc_link': full_doc_link,
-                        'is_open_access': is_open_access,
-                    }, ignore_index=True)
-            else:
-                get_paper_info_result = 'no_code'
-            error_traceback = ""
-        except Exception as err:
-            get_paper_info_result = 'exception'
-            error_traceback = traceback.format_exc()
-
-        scrape_time = time.time() - start_time
-
-        status_df = status_df.append({
-            'url': url,
-            'literature_site': literature_site,
-            'get_paper_info_result': get_paper_info_result,
-            'title_len': len(title) if isinstance(title,str) else 0,
-            'abstract_len': len(abstract) if isinstance(abstract,str) else 0,
-            'doi_len': len(doi) if isinstance(doi,str) else 0,
-            'full_doc_link_len': len(full_doc_link) if isinstance(full_doc_link,str) else 0,
-            'is_open_access': is_open_access,
-            'num_labels': len(labels),
-            'error_traceback': error_traceback,
-            'scrape_time': scrape_time,
-        }, ignore_index=True, na_filter= False)
-
-    return transformed_df, status_df
-
-def load(df, csv_file):
-    df.to_csv(csv_file)
-
-
-def transformed_data_check(df):
-    print('Number of empty cells in the columns')
-    for name in df.columns:
-        c = (df[name] == '').sum()
-        print(f'{name}: {c}')
-
-    with pd.option_context('display.max_rows', None,
-                           'display.max_columns', None,
-                           'display.max_colwidth', 100):
-        print(df.describe())
-
-# def filter_based_on_what_already_has_been_done(input_df, output_csv):
-#     # does the output csv exist? I not, no filtering
-#     if not Path(output_csv).exists():
-#         return input_df
-#     df_exists = pd.read_csv(output_csv)
+# Not used currently
+# def transform(df):
+#     # Make an empty table for the results of the transform
+#     transformed_df = pd.DataFrame(columns=['title', 'doi', 'abstract', 'labels', 'url',
+#                                            'literature_site',
+#                                            'full_doc_link', 'is_open_access'])
 #
-#     # make a new empty one
-#     new_df = pd.DataFrame(columns=input_df.columns)
+#     # Need to keep track of the status of each attempt to get paper info
+#     status_df = pd.DataFrame(columns=['url', 'literature_site', 'get_paper_info_result',
+#                                       'title_len', 'abstract_len', 'doi_len',
+#                                       'pdf_len', 'is_open_access',
+#                                       'num_labels',
+#                                       'error_traceback', 'scrape_time'])
+#     status_df.astype(int)  # No floats
 #
-#     # Loop through the input df
-#     for index, row in input_df.iterrows():
-#         _, _, journal, url, labels, level_2, level_3_new, level_3_old, link_press, doi, title, abstract = row
+#     # Loop through the records to get paper info
+#     for index, row in df[['Primary lit site', 'Functions Level I','Abstract']].iterrows():
+#         url, labels, abstract = row
+#         # if pd.isnull(labels):
+#         #     labels = ''
+#         print(f"{index} url: {url}")
 #
-#         # check to see if the url matches one in the existing df
-#         matches = df_exists.loc[df_exists['url'] == url]
+#         start_time = time.time()
 #
-#         if not matches.empty :
-#             if len(matches) > 1:
-#                 raise(ValueError, "Do not know what to do with more than one match")
-#             # check to see if we have good values
-#             # ~~~~~~~~!!!!!!!!!!!!!! need to handle nan's also
-#             if len(matches['doi']) and len(matches['title']) and len(matches['abstract']) and len(matches['full_doc_link']):
-#                 continue
+#         literature_site = which_literature_site(url)
 #
-#         # Otherwise add it to the list of pages to scrape
-#         new_df = new_df.append(row,ignore_index=True)
+#         # continue
+#         title = doi = abstract = full_doc_link = ''
+#         is_open_access = False
 #
-#     return new_df
+#         # fix labels
+#         labels = labels_fix(labels)
+#
+#         try:
+#             paper_info = get_paper_info(url)
+#             if paper_info:
+#                 title, doi, abstract, full_doc_link, is_open_access, is_blocked = paper_info
+#                 if is_blocked:
+#                     get_paper_info_result = 'blocked'
+#                 else:
+#                     get_paper_info_result = 'no_exception'
+#
+#                     abstract = abstract_fix(abstract)
+#
+#                     transformed_df = transformed_df.append({
+#                         'title': title,
+#                         'doi': doi,
+#                         'abstract': abstract,
+#                         'labels': labels,
+#                         'url': url,
+#                         'literature_site': literature_site,
+#                         'full_doc_link': full_doc_link,
+#                         'is_open_access': is_open_access,
+#                     }, ignore_index=True)
+#             else:
+#                 get_paper_info_result = 'no_code'
+#             error_traceback = ""
+#         except Exception as err:
+#             get_paper_info_result = 'exception'
+#             error_traceback = traceback.format_exc()
+#
+#         scrape_time = time.time() - start_time
+#
+#         status_df = status_df.append({
+#             'url': url,
+#             'literature_site': literature_site,
+#             'get_paper_info_result': get_paper_info_result,
+#             'title_len': len(title) if isinstance(title,str) else 0,
+#             'abstract_len': len(abstract) if isinstance(abstract,str) else 0,
+#             'doi_len': len(doi) if isinstance(doi,str) else 0,
+#             'full_doc_link_len': len(full_doc_link) if isinstance(full_doc_link,str) else 0,
+#             'is_open_access': is_open_access,
+#             'num_labels': len(labels),
+#             'error_traceback': error_traceback,
+#             'scrape_time': scrape_time,
+#         }, ignore_index=True, na_filter= False)
+#
+#     return transformed_df, status_df
 
+# def load(df, csv_file):
+#     df.to_csv(csv_file)
 
-def scrape_paper_info(input_record, df_status):
-    url = input_record['url']
-    print(f"url: {url}")
+# def transformed_data_check(df):
+#     print('Number of empty cells in the columns')
+#     for name in df.columns:
+#         c = (df[name] == '').sum()
+#         print(f'{name}: {c}')
+#
+#     with pd.option_context('display.max_rows', None,
+#                            'display.max_columns', None,
+#                            'display.max_colwidth', 100):
+#         print(df.describe())
 
-    start_time = time.time()
-
-    # fix labels
-    input_record['label_level_1'] = labels_fix(input_record['label_level_1'])
-    input_record['label_level_2'] = labels_fix(input_record['label_level_2'])
-    input_record['label_level_3'] = labels_fix(input_record['label_level_3'])
-
-    title = doi = abstract = full_doc_link = ''
-    is_open_access = False
-
-    try:
-        paper_info = get_paper_info(url)
-        if paper_info:
-            title, doi, abstract, full_doc_link, is_open_access, is_blocked = paper_info
-
-            # fix abstract
-            abstract = abstract_fix(abstract)
-
-            get_paper_info_result = 'no_exception'
-
-            input_record['title'] = title
-            input_record['doi'] = doi
-            input_record['abstract'] = abstract
-            input_record['full_doc_link'] = full_doc_link
-            input_record['is_open_access'] = is_open_access
-        else:
-            get_paper_info_result = 'no_code'
-        error_traceback = ""
-    except Exception as err:
-        get_paper_info_result = 'exception'
-        error_traceback = traceback.format_exc()
-
-    scrape_time = time.time() - start_time
-
-    literature_site = which_literature_site(url)
-    input_record['literature_site'] = literature_site
-
-    df_status = df_status.append({
-        'url': url,
-        'literature_site': literature_site,
-        'get_paper_info_result': get_paper_info_result,
-        'title_len': len(title) if isinstance(title, str) else 0,
-        'abstract_len': len(abstract) if isinstance(abstract, str) else 0,
-        'doi_len': len(doi) if isinstance(doi, str) else 0,
-        'full_doc_link_len': len(full_doc_link) if isinstance(full_doc_link, str) else 0,
-        'is_open_access': is_open_access,
-        'num_labels': len(input_record['label_level_1']),
-        'error_traceback': error_traceback,
-        'scrape_time': scrape_time,
-    }, ignore_index=True)
-
-    return input_record, df_status
 
 def update_and_append_paper_info_table(df_input, df_paper_info_db):
 
@@ -238,12 +161,9 @@ def update_and_append_paper_info_table(df_input, df_paper_info_db):
     df_status.astype(int)  # No floats
 
     for index, row in df_input.iterrows():
-        # input_record = dict(zip(standard_columns, row))
         input_record = row.to_dict()
 
         print(f"{index}: input_record['url']: {input_record['url']}")
-        # look in df_output for a matching doi
-        #    !!!! comes back as Nan
         matching_record = df_paper_info_db.loc[df_paper_info_db['url'] == input_record['url']]
 
         if not matching_record.empty :
@@ -270,9 +190,7 @@ def update_and_append_paper_info_table(df_input, df_paper_info_db):
                     if matching_record[key] != input_record[key]:
                         print(f"Warning! For {input_record['url']} the {key} value changed.")
                         matching_record[key] = input_record[key]
-            # replace?
             print("updating")
-            # df_paper_info_db.loc[df_paper_info_db['url'] == input_record['url']] = matching_record
             i = df_paper_info_db[ df_paper_info_db['url'] == input_record['url']].index
             df_paper_info_db.drop(i, inplace=True)
             df_paper_info_db = df_paper_info_db.append(input_record, ignore_index=True)
@@ -291,23 +209,16 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(prog = sys.argv[0],
                                      description = "scape sites and generate csv with info.")
 
-    # usually "../data/Colleen_and_Alex.csv"
     parser.add_argument('input_csv', type=str, help='input CSV file that needs additional info')
-    # usually "Colleen_and_Alex_transformed.csv"
     parser.add_argument('output_csv', type=str, help='output CSV file')
-    # usually "Colleen_and_Alex_etl_status.csv"
     parser.add_argument('status_csv', type=str, help='status CSV file')
-
     parser.add_argument('-n', help='limit the number of journals to this number',
                         default=None, type=int)
-
     parser.add_argument('--filter', type=str,
                         help='filter based on matching this search string in Primary Lit Site',
                         default=None)
-
     parser.add_argument("--env_path", help = "path to .env file containing API keys",
                         default = default_path_to_env, type = str)
-
     parser.add_argument('--new_db', action='store_true', default=False, dest='new_db',
                         help="create new empty db csv file.")
 
@@ -319,19 +230,21 @@ if __name__ == "__main__":
 
     # raw_data_check(df)
 
+    # These allow you to limit what gets scraped.
+    # The filter lets you limit scraping to a certain journal site
+    # The n value lets you limit the number of papers scraped. Good for debugging.
     if args.filter:
         df = filter_by_lit_site(df, args.filter)
-
-    # # Only scrape what needs to be scraped. First see what is in the output csv
-    # df = filter_based_on_what_already_has_been_done(df, args.output_csv)
-
     if args.n:
         df = filter_by_count(df, args.n)
 
+    # You have the option to create a completely new database rather than starting
+    #  from an existing one and updating it
     if args.new_db:
         df_paper_info_db = pd.DataFrame(columns=standard_columns +['literature_site'])
     else:
-        # Make a timestamped backup copy
+        # Make a timestamped backup copy of the primary CSV database file in case something
+        #   bad happens
         path = Path(args.output_csv)
 
         mtime = datetime.datetime.fromtimestamp(path.stat().st_mtime)
@@ -348,10 +261,4 @@ if __name__ == "__main__":
 
     df_paper_info_db.to_csv(args.output_csv, index=False)
 
-    # transformed_df, status_df = transform(df)
-    #
-    # transformed_data_check(transformed_df)
-    #
-    # load(transformed_df, args.output_csv)
-    #
     save_status(df_status, args.status_csv)
